Added PlayerMoves class
The PlayerMoves class contains all the types of moves that a non-human Player can make. PlayerMoves can make a random move, the best move, or the worst move. Two private attributes are used: _othello is the current Othello game stat and _character is the player that is making the move. Together these attributes are used to determine possible moves. _othello is used to create copies of the original game state and run simulations of whether _character can place a token on a given spot and how many tokens _character has on the board.