Added PlayerEasy and PlayerMedium classes
Non-human Players are separated from the move functions because it will be easier to add and adjust difficulties across all non-human players. PlayerEasy and PlayerMedium are non-human Players. When getMove is called upon, it will return a Move given their difficulty. PlayerEasy will return worst_move in the PlayerMoves class and PlayerMedium will return random_move in the PlayerMoves class.