Made GUI Application Run
The Application.py file is what needs to be run in order to see the GUI.
It makes the Tk(), and adds the significant frames into the root.
The individual frames are constructed using some other classes which
make the widgets and store it into their frames, these classes are,
ImageFrame and ButtonFrame in the Widgets package.
A Token class is created to act as a Button but provide us the
(x,y) coordinate of where it is located so that it can be linked to the
board in the future.
Added a visual representation of the board, which is an 8 by 8 grid of
buttons, this was made through the ImageFrame class.
Added buttons for the Settings, Game Modes, and Quit options to be
implemented in the future.
Waiting on Othello model functionality to be completed so that the
GUI (Application.py) can be linked to the model.