Create play method for controller superclass
The play method is responsible for continuously
getting move from a player and making a move on
the board until the game ends. For testing purposes,
there will be reports made to the console on various
game information every time a move is made and when
the game is over.